requirements.txt
vastai-sdk>=0.3.0
Pillow
